refactor: FindGlossaryThreadTest
- Move test cases to GlossarySearcherTest.java
