+++ test/src/org/omegat/gui/glossary/GlossarySearcherTest.java
@@ -29,6 +29,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -40,11 +41,13 @@ import org.omegat.core.data.EntryKey;
 import org.omegat.core.data.NotLoadedProject;
 import org.omegat.core.data.ProjectProperties;
 import org.omegat.core.data.SourceTextEntry;
+import org.omegat.tokenizer.DefaultTokenizer;
 import org.omegat.tokenizer.ITokenizer;
 import org.omegat.tokenizer.LuceneCJKTokenizer;
 import org.omegat.tokenizer.LuceneEnglishTokenizer;
 import org.omegat.tokenizer.LuceneJapaneseTokenizer;
 import org.omegat.util.Language;
+import org.omegat.util.Preferences;
 
 /**
  * @author Hiroshi Miura
@@ -56,10 +59,11 @@ public class GlossarySearcherTest extends TestCore {
         String translationText = "translation";
         String commentText = "comment";
         ITokenizer tok = new LuceneEnglishTokenizer();
-        Language language = new Language("en");
-        setupProject(language);
+        Language srcLang = new Language("en");
+        Language trLang = new Language("de");
+        setupProject(srcLang);
         List<GlossaryEntry> entries = Arrays.asList(new GlossaryEntry(sourceText, translationText, commentText, true, "origin"));
-        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, language, entries);
+        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, srcLang, trLang, entries);
         assertEquals(1, result.size());
         assertEquals(sourceText, result.get(0).getSrcText());
         assertEquals(commentText, result.get(0).getCommentText());
@@ -85,10 +89,11 @@ public class GlossarySearcherTest extends TestCore {
         String commentText = "comment";
         ITokenizer tok = new LuceneCJKTokenizer();
         Language language = new Language("ko");
+        Language trLang = new Language("en");
         setupProject(language);
         List<GlossaryEntry> entries = Collections.singletonList(new GlossaryEntry(sourceText,
                 translationText, commentText, true, "origin"));
-        List<GlossaryEntry> result = glossarySearcherCommon(segmentText, tok, language, entries);
+        List<GlossaryEntry> result = glossarySearcherCommon(segmentText, tok, language, trLang, entries);
         assertEquals(1, result.size());
     }
 
@@ -99,9 +104,10 @@ public class GlossarySearcherTest extends TestCore {
         String commentText = "comment";
         ITokenizer tok = new LuceneJapaneseTokenizer();
         Language language = new Language("ja");
+        Language trLang = new Language("en");
         setupProject(language);
         List<GlossaryEntry> entries = Arrays.asList(new GlossaryEntry(sourceText, translationText, commentText, true, "origin"));
-        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, language, entries);
+        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, language, trLang, entries);
         assertEquals(1, result.size());
         assertEquals(sourceText, result.get(0).getSrcText());
         assertEquals(commentText, result.get(0).getCommentText());
@@ -112,16 +118,18 @@ public class GlossarySearcherTest extends TestCore {
     public void testGlossarySearcherJapanese2() {
         String sourceText = "\u5834\u6240";
         Language language = new Language("ja");
+        Language trLang = new Language("en");
         setupProject(language);
         ITokenizer tok = new LuceneJapaneseTokenizer();
         List<GlossaryEntry> entries = Arrays.asList(new GlossaryEntry("\u5857\u5E03", "wrong", "", true, "origin"));
-        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, language, entries);
+        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, language, trLang, entries);
         assertEquals(0, result.size());
     }
 
     @Test
     public void testGlossarySearcherJapaneseLongText() {
         Language language = new Language("ja");
+        Language trLang = new Language("en");
         setupProject(language);
         ITokenizer tok = new LuceneJapaneseTokenizer();
         List<GlossaryEntry> entries = Arrays.asList(
@@ -131,8 +139,93 @@ public class GlossarySearcherTest extends TestCore {
                 new GlossaryEntry("\u5730\u57DF\u5316", "localization", "", true, "")
         );
         String sourceText = "OmegaT\u306E\u30E6\u30FC\u30B6\u30FC\u30A4\u30F3\u30BF\u30FC\u30D5\u30A7\u30FC\u30B9\u3084\u30D8\u30EB\u30D7\u30C6\u30AD\u30B9\u30C8\u3092\u3001\u3055\u307E\u3056\u307E\u306A\u8A00\u8A9E\u3078\u7FFB\u8A33\u3057\u3066\u304F\u3060\u3055\u3063\u305F\u65B9\u3005\u306B\u611F\u8B1D\u3057\u307E\u3059\u3002\u305D\u3057\u3066\u3001\u7FFB\u8A33\u304C\u306A\u3055\u308C\u3066\u3044\u306A\u3044\u8A00\u8A9E\u304C\u307E\u3060\u6570\u5343\u6B8B\u3063\u3066\u3044\u307E\u3059\uFF01OmegaT \u306E\u591A\u8A00\u8A9E\u3078\u306E\u5730\u57DF\u5316\u306F\u3001\u6301\u7D9A\u7684\u306A\u4F5C\u696D\u3067\u3082\u3042\u308A\u307E\u3059\u3002\u306A\u305C\u306A\u3089\u3001\u65B0\u3057\u3044\u6A5F\u80FD\u304C\u7D76\u3048\u305A\u8FFD\u52A0\u3055\u308C\u3066\u3044\u308B\u304B\u3089\u3067\u3059\u3002OmegaT\u306E\u30ED\u30FC\u30AB\u30E9\u30A4\u30BA/\u7FFB\u8A33\u306B\u95A2\u3059\u308B\u8A73\u7D30\u306B\u3064\u3044\u3066\u306F\u3001OmegaT\u30ED\u30FC\u30AB\u30EA\u30BC\u30FC\u30B7\u30E7\u30F3\u30B3\u30FC\u30C7\u30A3\u30CD\u30FC\u30BF\u30FC\u306B\u304A\u554F\u3044\u5408\u308F\u305B\u304F\u3060\u3055\u3044\u3002";
-        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, language, entries);
+        List<GlossaryEntry> result = glossarySearcherCommon(sourceText, tok, language, trLang, entries);
         assertEquals(3, result.size());
+    }
+
+    @Test
+    public void testEntriesSortEn() {
+        Language srcLang = new Language("en_US");
+        Language targetLang = new Language("en_GB");
+        ITokenizer tok = new DefaultTokenizer();
+        GlossarySearcher searcher = new GlossarySearcher(tok, srcLang, targetLang, false);
+        List<GlossaryEntry> entries = new ArrayList<>();
+        entries.add(new GlossaryEntry("dog", "doggy", "cdog", false, null));
+        entries.add(new GlossaryEntry("cat", "catty", "ccat", false, null));
+        entries.add(new GlossaryEntry("cat", "mikeneko", "ccat", false, null));
+        entries.add(new GlossaryEntry("zzz", "zzz", "czzz", true, null));
+        entries.add(new GlossaryEntry("horse", "catty", "chorse", false, null));
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_LENGTH, true);
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_SRC_LENGTH, false);
+        searcher.sortGlossaryEntries(entries);
+        assertEquals("zzz", entries.get(0).getSrcText());
+        assertEquals("cat", entries.get(1).getSrcText());
+        assertEquals("mikeneko", entries.get(1).getLocText());
+        assertEquals("cat", entries.get(2).getSrcText());
+        assertEquals("catty", entries.get(2).getLocText());
+        assertEquals("dog", entries.get(3).getSrcText());
+        assertEquals("horse", entries.get(4).getSrcText());
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_LENGTH, false);
+        searcher.sortGlossaryEntries(entries);
+        assertEquals("zzz", entries.get(0).getSrcText());
+        assertEquals("cat", entries.get(1).getSrcText());
+        assertEquals("catty", entries.get(1).getLocText());
+        assertEquals("cat", entries.get(2).getSrcText());
+        assertEquals("mikeneko", entries.get(2).getLocText());
+        assertEquals("dog", entries.get(3).getSrcText());
+        assertEquals("horse", entries.get(4).getSrcText());
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_SRC_LENGTH, true);
+        searcher.sortGlossaryEntries(entries);
+        assertEquals("zzz", entries.get(0).getSrcText());
+        assertEquals("cat", entries.get(1).getSrcText());
+        assertEquals("catty", entries.get(1).getLocText());
+        assertEquals("cat", entries.get(2).getSrcText());
+        assertEquals("mikeneko", entries.get(2).getLocText());
+        assertEquals("dog", entries.get(3).getSrcText());
+        assertEquals("horse", entries.get(4).getSrcText());
+    }
+
+    @Test
+    public void testEntriesSortJA() {
+        Language lang = new Language("ja_JP");
+        Language targetLang = new Language("en_GB");
+        ITokenizer tok = new DefaultTokenizer();
+        GlossarySearcher searcher = new GlossarySearcher(tok, lang, targetLang, false);
+        List<GlossaryEntry> entries = new ArrayList<>();
+        entries.add(new GlossaryEntry("向上", "enhance", "", false, null));
+        entries.add(new GlossaryEntry("向", "direct", "", false, null));
+        entries.add(new GlossaryEntry("上", "on", "", false, null));
+        entries.add(new GlossaryEntry("上", "up to", "", false, null));
+        entries.add(new GlossaryEntry("トヨタ自動車", "toyota motors", "", false, null));
+        entries.add(new GlossaryEntry("トヨタ", "toyota", "", false, null));
+        entries.add(new GlossaryEntry("さくら", "cherry blossom", "", false, null));
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_LENGTH, true);
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_SRC_LENGTH, false);
+        searcher.sortGlossaryEntries(entries);
+        assertEquals("さくら", entries.get(0).getSrcText());
+        assertEquals("トヨタ", entries.get(1).getSrcText());
+        assertEquals("トヨタ自動車", entries.get(2).getSrcText());
+        assertEquals("向", entries.get(3).getSrcText());
+        assertEquals("向上", entries.get(4).getSrcText());
+        assertEquals("up to", entries.get(5).getLocText());
+        assertEquals("on", entries.get(6).getLocText());
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_LENGTH, false);
+        searcher.sortGlossaryEntries(entries);
+        assertEquals("cherry blossom", entries.get(0).getLocText());
+        assertEquals("toyota", entries.get(1).getLocText());
+        assertEquals("toyota motors", entries.get(2).getLocText());
+        assertEquals("direct", entries.get(3).getLocText());
+        assertEquals("enhance", entries.get(4).getLocText());
+        assertEquals("on", entries.get(5).getLocText());
+        assertEquals("up to", entries.get(6).getLocText());
+        Preferences.setPreference(Preferences.GLOSSARY_SORT_BY_SRC_LENGTH, true);
+        searcher.sortGlossaryEntries(entries);
+        assertEquals("toyota motors", entries.get(1).getLocText());
+        assertEquals("toyota", entries.get(2).getLocText());
+        assertEquals("enhance", entries.get(3).getLocText());
+        assertEquals("direct", entries.get(4).getLocText());
+        assertEquals("on", entries.get(5).getLocText());
+        assertEquals("up to", entries.get(6).getLocText());
     }
 
     private void setupProject(Language language) {
@@ -162,11 +255,11 @@ public class GlossarySearcherTest extends TestCore {
         });
     }
 
-    private List<GlossaryEntry> glossarySearcherCommon(String sourceText, ITokenizer tok, Language language,
-                                                       List<GlossaryEntry> entries) {
+    private List<GlossaryEntry> glossarySearcherCommon(String sourceText, ITokenizer tok, Language srcLang,
+                                                       Language trLang, List<GlossaryEntry> entries) {
         EntryKey key = new EntryKey("file", sourceText, "id", "prev", "next", "path");
         SourceTextEntry ste = new SourceTextEntry(key, 1, new String[0], sourceText, Collections.emptyList());
-        GlossarySearcher searcher = new GlossarySearcher(tok, language, false);
+        GlossarySearcher searcher = new GlossarySearcher(tok, srcLang, trLang,false);
         return searcher.searchSourceMatches(ste, entries);
     }
 }
